+++ src/Functions/Functions.php
@@ -7,6 +7,7 @@ use Katrina\Functions\Traits\{
     AggregateFunctionsTrait,
     MathFunctionsTrait
 };
+use Symfony\Component\Uid\Uuid;
 
 class Functions
 {
@@ -29,7 +30,6 @@ class Functions
     public static function subquery(string $query): string
     {
         self::$query = $query;
-
         return self::$query;
     }
 
@@ -39,5 +39,27 @@ class Functions
     public static function getQuery(): string
     {
         return self::$query;
+    }
+
+    /**
+     * Generate a raw binary string
+     *
+     * @return string
+     */
+    public static function uuidToBin(): string
+    {
+        return Uuid::v4()->toBinary();
+    }
+
+    /**
+     * Returns the identifier as a RFC4122 case insensitive string
+     *
+     * @param string $binary_uuid
+     * 
+     * @return string
+     */
+    public static function binToUuid(string $binary_uuid): string
+    {
+        return Uuid::fromBinary($binary_uuid)->toRfc4122();
     }
 }